feat: indicate slow internet connection in convo list
