+++ src/script/util/ConnectionQualityHandler/ConnectionQualityHandler.ts
@@ -17,8 +17,6 @@
  *
  */
 
-import {useEffect, useState} from 'react';
-
 import {TimeInMillis} from '@wireapp/commons/lib/util/TimeUtil';
 
 import {getLogger} from 'Util/Logger';
@@ -78,22 +76,4 @@ export const getConnectionQualityHander = (): ConnectionQualityHandler | null =>
     refresh: (callback: (isSlow: boolean) => void) => onChange(callback),
     subscribe,
   };
-};
-
-export const useConnectionQuality = () => {
-  const [isSlow, setIsSlow] = useState(false);
-
-  useEffect(() => {
-    const connectionQualityHandler = getConnectionQualityHander();
-
-    if (!connectionQualityHandler) {
-      return;
-    }
-
-    const unsubscribe = connectionQualityHandler.subscribe(setIsSlow);
-
-    return unsubscribe;
-  }, []);
-
-  return {isSlow};
 };
